Move everything to MultiWriter and use select_all
Still haven't figured out how to create copy futures now that the future
doesn't own the read/write half types, which is a little strange tbh.
I'm sure it's a simple rust-specific problem but I'm struggling to
figure out how to address this without lots and lots of boilerplate
code.