New design: use channels to copy messages to different spawns
Then we can try_join on different copy futures to keep the read/write
half on the same stack.

Remaining: Currently MultiWriter is set to work with an explicit type
(currently showing up as WriteHalf), we need to change this to a dyn
AsyncWrite while keeping the select_all future happy.